+++ shiori/api/views.py
@@ -5,20 +5,27 @@ from django.contrib.auth.models import AnonymousUser
 from rest_framework import viewsets
 from rest_framework.response import Response
 from rest_framework.permissions import (IsAuthenticated,
-                                        IsAuthenticatedOrReadOnly)
+                                        IsAuthenticatedOrReadOnly,
+                                        IsAdminUser)
 from rest_framework.exceptions import NotAuthenticated
-from shiori.bookmark.models import Category, Tag, Bookmark, BookmarkTag
-from shiori.api.permissions import IsOwnerOrReadOnly
+from shiori.bookmark.models import (Category,
+                                    Tag,
+                                    Bookmark,
+                                    BookmarkTag,
+                                    FeedSubscription)
+from shiori.api.permissions import (IsOwnerOrReadOnly,
+                                    IsAuthenticatedAndCreateReadOnly)
 from shiori.api.serializers import (CategorySerializer,
                                     TagSerializer,
                                     BookmarkSerializer,
-                                    BookmarkTagSerializer)
+                                    BookmarkTagSerializer,
+                                    FeedSubscriptionSerializer)
 
 
 class CategoryViewSet(viewsets.ModelViewSet):
     queryset = Category.objects.all()
     serializer_class = CategorySerializer
-    permission_classes = (IsAuthenticatedOrReadOnly,)
+    permission_classes = (IsAuthenticatedAndCreateReadOnly,)
 
     def pre_save(self, obj):
         obj.category = escape(self.request.DATA.get('category'))
@@ -27,7 +34,7 @@ class CategoryViewSet(viewsets.ModelViewSet):
 class TagViewSet(viewsets.ModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
-    permission_classes = (IsAuthenticatedOrReadOnly,)
+    permission_classes = (IsAuthenticatedAndCreateReadOnly,)
     paginate_by = 50
 
     def pre_save(self, obj):
@@ -63,3 +70,18 @@ class BookmarkTagViewSet(viewsets.ModelViewSet):
     queryset = BookmarkTag.objects.all()
     serializer_class = BookmarkTagSerializer
     permission_classes = (IsAuthenticatedOrReadOnly,)
+
+
+class FeedSubscriptionViewSet(viewsets.ModelViewSet):
+    queryset = FeedSubscription.objects.all()
+    serializer_class = FeedSubscriptionSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def pre_save(self, obj):
+        if isinstance(self.request.user, AnonymousUser):
+            raise NotAuthenticated
+        else:
+            obj.owner = self.request.user
+
+    def get_queryset(self):
+        return FeedSubscription.objects.filter(owner=self.request.user)