+++ shiori/core/urls.py
@@ -10,6 +10,6 @@ urlpatterns = patterns('',
                        (r'^$', RedirectView.as_view(url='/shiori/')),
                        (r'^logout/$', 'django.contrib.auth.views.logout',
                         {'template_name': 'bookmark/logout.html'}),
-                       (r'^v1/', include('shiori.api.urls')),
+                       (r'^api/', include('shiori.api.urls')),
                        (r'^shiori/', include('shiori.bookmark.urls')),
                        url(r'^admin/', include(admin.site.urls)),)