+++ shiori_tests/settings.py
@@ -107,11 +107,11 @@ REST_FRAMEWORK = {
         'rest_framework.authentication.SessionAuthentication',
         ),
     'DEFAULT_PERMISSION_CLASSES': [
-        #'rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly',
-        #'rest_framework.permissions.DjangoModelPermissions',
-        #'rest_framework.permissions.AllowAny',
+        # 'rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly',
+        # 'rest_framework.permissions.DjangoModelPermissions',
+        # 'rest_framework.permissions.AllowAny',
         'rest_framework.permissions.IsAuthenticated',
-        #'rest_framework.permissions.IsAdminUser',
+        # 'rest_framework.permissions.IsAdminUser',
         ],
     'PAGINATE_BY': 20}
 
@@ -129,3 +129,9 @@ OPENID_USE_EMAIL_FOR_USERNAME = True
 # Django 1.6 can't serialize openid.yadis.manager.YadisServiceManager objects
 # http://bazaar.launchpad.net/~strycore/django-openid-auth/trunk/revision/115
 SESSION_SERIALIZER = 'django.contrib.sessions.serializers.PickleSerializer'
+
+
+# loopback addresses have been excluded in default.
+FEED_EXCLUDE_FQDN = (
+    # 'shiori.example.org',
+)