+++ shiori_tests/tests/test_api.py
@@ -14,20 +14,21 @@ class APITest(TestCase):
     def setUp(self):
         User.objects.create_user(v.username0, v.email0, v.password0)
         User.objects.create_user(v.username1, v.email1, v.password1)
+        User.objects.create_superuser(v.username2, v.email2, v.password2)
         self.client.login(username=v.username0, password=v.password0)
 
         payload = {'category': v.category0}
-        self.client.post('/v1/categories',
+        self.client.post('/api/categories',
                          content_type='application/json',
                          data=json.dumps(payload))
 
         payload = {'tag': v.tag0}
-        self.client.post('/v1/tags',
+        self.client.post('/api/tags',
                          content_type='application/json',
                          data=json.dumps(payload))
 
         payload = {'tag': v.tag1}
-        self.client.post('/v1/tags',
+        self.client.post('/api/tags',
                          content_type='application/json',
                          data=json.dumps(payload))
 
@@ -36,7 +37,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description0,
                    'is_hide': False}
-        self.client.post('/v1/bookmarks',
+        self.client.post('/api/bookmarks',
                          content_type='application/json',
                          data=json.dumps(payload))
         payload = {'url': v.url1,
@@ -44,48 +45,48 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description1,
                    'is_hide': True}
-        self.client.post('/v1/bookmarks',
+        self.client.post('/api/bookmarks',
                          content_type='application/json',
                          data=json.dumps(payload))
         payload = {'bookmark': v.url0, 'tag': v.tag0}
-        self.client.post('/v1/bookmark_tags',
+        self.client.post('/api/bookmark_tags',
                          content_type='application/json',
                          data=json.dumps(payload))
 
     def test_api_categories_url_resolve(self):
-        response = self.client.get('/v1/categories')
+        response = self.client.get('/api/categories')
         self.assertContains(response,
                             'results',
                             status_code=200)
 
     def test_api_tags_url_resolve(self):
-        response = self.client.get('/v1/tags')
+        response = self.client.get('/api/tags')
         self.assertContains(response,
                             'results',
                             status_code=200)
 
     def test_api_bookmarks_url_resolve(self):
-        response = self.client.get('/v1/bookmarks')
+        response = self.client.get('/api/bookmarks')
         self.assertContains(response,
                             'results',
                             status_code=200)
 
     def test_api_bookmark_tags_url_resolve(self):
-        response = self.client.get('/v1/bookmark_tags')
+        response = self.client.get('/api/bookmark_tags')
         self.assertContains(response,
                             'results',
                             status_code=200)
 
     def test_post_tag(self):
         payload = {'tag': v.tag2}
-        response = self.client.post('/v1/tags',
+        response = self.client.post('/api/tags',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 201)
 
     def test_post_tag_conflict(self):
         payload = {'tag': v.tag0}
-        response = self.client.post('/v1/tags',
+        response = self.client.post('/api/tags',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -93,45 +94,46 @@ class APITest(TestCase):
                             status_code=400)
 
     def test_get_tags(self):
-        response = self.client.get('/v1/tags')
+        response = self.client.get('/api/tags')
         self.assertContains(response,
                             '"tag": "%s"' % v.tag0,
                             status_code=200)
 
     def test_get_tag(self):
-        r = self.client.get('/v1/tags')
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.get('/v1/tags/%s' % id)
+        response = self.client.get('/api/tags/%s' % id)
         self.assertContains(response,
                             '"tag": "%s"' % v.tag0,
                             status_code=200)
 
-    def test_delete_tag(self):
-        r = self.client.get('/v1/tags')
+    def test_cannot_delete_tag(self):
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/tags/%s' % id)
-        self.assertEqual(response.status_code, 204)
+        response = self.client.delete('/api/tags/%s' % id)
+        self.assertEqual(response.status_code, 403)
 
-    def test_put_tag(self):
-        r = self.client.get('/v1/tags')
+    def test_cannot_put_tag(self):
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'tag': v.tag2}
-        response = self.client.put('/v1/tags/%s' % id,
+        response = self.client.put('/api/tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertContains(response,
-                            '"tag": "%s"' % v.tag2,
-                            status_code=200)
+                            '{"detail": "You do not have permission to'
+                            ' perform this action."}',
+                            status_code=403)
 
     def test_get_tags_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/tags')
+        r = self.client.get('/api/tags')
         self.assertContains(r,
                             'results',
                             status_code=200)
 
         id = json.loads(r.content).get('results')[0].get('id').encode('utf-8')
-        response = self.client.get('/v1/tags/%s' % id)
+        response = self.client.get('/api/tags/%s' % id)
         self.assertContains(response,
                             '"tag": "%s"' % v.tag0,
                             status_code=200)
@@ -139,7 +141,7 @@ class APITest(TestCase):
     def test_post_tags_by_anonymous(self):
         self.client.logout()
         payload = {'tag': v.tag1}
-        response = self.client.post('/v1/tags',
+        response = self.client.post('/api/tags',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
@@ -147,47 +149,45 @@ class APITest(TestCase):
     def test_delete_tag_by_anonymous(self):
         self.client.logout()
 
-        r = self.client.get('/v1/tags')
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/tags/%s' % id)
+        response = self.client.delete('/api/tags/%s' % id)
         self.assertEqual(response.status_code, 403)
 
     def test_put_tag_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/tags')
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'tag': v.tag1}
-        response = self.client.put('/v1/tags/%s' % id,
+        response = self.client.put('/api/tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
 
-    # ToDo: delete is allowed by admin user only.
-    @unittest.skip("ToDo skipping")
-    def test_delete_tag_by_another_user(self):
+    # [Fixed] ToDo: delete is allowed by admin user only.
+    def test_delete_tag_by_superuser(self):
         self.client.logout()
-        self.client.login(username=v.username1, password=v.password1)
-        r = self.client.get('/v1/tags')
+        self.client.login(username=v.username2, password=v.password2)
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/tags/%s' % id)
-        self.assertEqual(response.status_code, 403)
+        response = self.client.delete('/api/tags/%s' % id)
+        self.assertEqual(response.status_code, 204)
 
-    # ToDo: put is allowed by admin user or anyone does not use.
-    @unittest.skip("ToDo skipping")
-    def test_put_tag_by_another_user(self):
+    # [Fixed] ToDo: put is allowed by admin user or anyone does not use.
+    def test_put_tag_by_superuser(self):
         self.client.logout()
-        self.client.login(username=v.username1, password=v.password1)
-        r = self.client.get('/v1/tags')
+        self.client.login(username=v.username2, password=v.password2)
+        r = self.client.get('/api/tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'tag': v.tag2}
-        response = self.client.put('/v1/tags/%s' % id,
+        response = self.client.put('/api/tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
-        self.assertEqual(response.status_code, 403)
+        self.assertEqual(response.status_code, 200)
 
     def test_post_category(self):
         payload = {'category': v.category1}
-        response = self.client.post('/v1/categories',
+        response = self.client.post('/api/categories',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -195,48 +195,49 @@ class APITest(TestCase):
                             status_code=201)
 
     def test_get_categories(self):
-        response = self.client.get('/v1/categories')
+        response = self.client.get('/api/categories')
         self.assertContains(response,
                             '"category": "%s"' % v.category0,
                             status_code=200)
 
     def test_get_category(self):
-        r = self.client.get('/v1/categories')
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.get('/v1/categories/%s' % id)
+        response = self.client.get('/api/categories/%s' % id)
         self.assertContains(response,
                             '"category": "%s"' % v.category0,
                             status_code=200)
 
-    def test_delete_category(self):
-        r = self.client.get('/v1/categories')
+    def test_cannot_delete_category(self):
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/categories/%s' % id)
-        self.assertEqual(response.status_code, 204)
+        response = self.client.delete('/api/categories/%s' % id)
+        self.assertEqual(response.status_code, 403)
 
-    def test_put_category(self):
-        r = self.client.get('/v1/categories')
+    def test_cannot_put_category(self):
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'category': v.category1}
-        response = self.client.put('/v1/categories/%s' % id,
+        response = self.client.put('/api/categories/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertContains(response,
-                            '"category":',
-                            status_code=200)
+                            '{"detail": "You do not have permission to'
+                            ' perform this action."}',
+                            status_code=403)
 
     def test_get_categories_by_anonymous(self):
         self.client.logout()
-        response = self.client.get('/v1/categories')
+        response = self.client.get('/api/categories')
         self.assertContains(response,
                             'results',
                             status_code=200)
 
     def test_get_category_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/categories')
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id').encode('utf-8')
-        response = self.client.get('/v1/categories/%s' % id)
+        response = self.client.get('/api/categories/%s' % id)
         self.assertContains(response,
                             '"category": "%s"' % v.category0,
                             status_code=200)
@@ -244,50 +245,48 @@ class APITest(TestCase):
     def test_post_category_by_anonymous(self):
         self.client.logout()
         payload = {'category': v.category1}
-        response = self.client.post('/v1/categories',
+        response = self.client.post('/api/categories',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
 
     def test_delete_category_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/categories')
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/categories/%s' % id)
+        response = self.client.delete('/api/categories/%s' % id)
         self.assertEqual(response.status_code, 403)
 
     def test_put_category_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/categories')
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'category': v.category1}
-        response = self.client.put('/v1/categories/%s' % id,
+        response = self.client.put('/api/categories/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
 
-    # ToDo: delete is allowed by admin user or anyone does not use.
-    @unittest.skip("ToDo skipping")
-    def test_delete_category_by_another_user(self):
+    # [Fixed] ToDo: delete is allowed by admin user or anyone does not use.
+    def test_delete_category_by_superuser(self):
         self.client.logout()
-        self.client.login(username=v.username1, password=v.password1)
-        r = self.client.get('/v1/categories')
+        self.client.login(username=v.username2, password=v.password2)
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/categories/%s' % id)
-        self.assertEqual(response.status_code, 403)
+        response = self.client.delete('/api/categories/%s' % id)
+        self.assertEqual(response.status_code, 204)
 
-    # ToDo: put is allowed by admin user or anyone does not use.
-    @unittest.skip("ToDo skipping")
-    def test_put_category_by_another_user(self):
+    # [Fixed] ToDo: put is allowed by admin user or anyone does not use.
+    def test_put_category_by_superuser(self):
         self.client.logout()
-        self.client.login(username=v.username1, password=v.password1)
-        r = self.client.get('/v1/categories')
+        self.client.login(username=v.username2, password=v.password2)
+        r = self.client.get('/api/categories')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id, 'category': v.category1}
-        response = self.client.put('/v1/categories/%s' % id,
+        response = self.client.put('/api/categories/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
-        self.assertEqual(response.status_code, 403)
+        self.assertEqual(response.status_code, 200)
 
     def test_post_bookmark_without_category(self):
         payload = {'url': v.url2,
@@ -295,7 +294,7 @@ class APITest(TestCase):
                    'category': '',
                    'description': v.description2,
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 400)
@@ -306,7 +305,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': None}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -319,7 +318,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 400)
@@ -330,7 +329,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 400)
@@ -341,7 +340,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': '',
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -354,7 +353,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -362,13 +361,13 @@ class APITest(TestCase):
                             status_code=201)
 
     def test_delete_bookmark(self):
-        r = self.client.get('/v1/bookmarks')
+        r = self.client.get('/api/bookmarks')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/bookmarks/%s' % id)
+        response = self.client.delete('/api/bookmarks/%s' % id)
         self.assertEqual(response.status_code, 204)
 
     def test_put_bookmark(self):
-        r = self.client.get('/v1/bookmarks')
+        r = self.client.get('/api/bookmarks')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id,
                    'url': v.url2,
@@ -376,7 +375,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': True}
-        response = self.client.put('/v1/bookmarks/%s' % id,
+        response = self.client.put('/api/bookmarks/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertContains(response,
@@ -387,7 +386,7 @@ class APITest(TestCase):
     def test_post_bookmark_tags(self):
         payload = {'bookmark': v.url0,
                    'tag': v.tag1}
-        response = self.client.post('/v1/bookmark_tags',
+        response = self.client.post('/api/bookmark_tags',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertContains(response,
@@ -395,18 +394,18 @@ class APITest(TestCase):
                             status_code=201)
 
     def test_delete_bookmark_tags(self):
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/bookmark_tags/%s' % id)
+        response = self.client.delete('/api/bookmark_tags/%s' % id)
         self.assertEqual(response.status_code, 204)
 
     def test_put_bookmark_tags(self):
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id,
                    'bookmark': v.url0,
                    'tag': v.tag1}
-        response = self.client.put('/v1/bookmark_tags/%s' % id,
+        response = self.client.put('/api/bookmark_tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertContains(response,
@@ -420,21 +419,21 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description2,
                    'is_hide': False}
-        response = self.client.post('/v1/bookmarks',
+        response = self.client.post('/api/bookmarks',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
 
     def test_delete_bookmark_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/bookmarks')
+        r = self.client.get('/api/bookmarks')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/bookmarks/%s' % id)
+        response = self.client.delete('/api/bookmarks/%s' % id)
         self.assertEqual(response.status_code, 403)
 
     def test_put_bookmark_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/bookmarks')
+        r = self.client.get('/api/bookmarks')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id,
                    'url': v.url0,
@@ -443,7 +442,7 @@ class APITest(TestCase):
                    'description': v.description1,
                    'is_hide': True}
 
-        response = self.client.put('/v1/bookmarks/%s' % id,
+        response = self.client.put('/api/bookmarks/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
@@ -452,12 +451,12 @@ class APITest(TestCase):
     def test_delete_bookmark_by_another_user(self):
         self.client.logout()
         self.client.login(username=v.username1, password=v.password1)
-        r = self.client.get('/v1/bookmarks?is_all=true')
+        r = self.client.get('/api/bookmarks?is_all=true')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/bookmarks/%s' % id)
+        response = self.client.delete('/api/bookmarks/%s' % id)
         self.assertEqual(response.status_code, 404)
 
-        response = self.client.delete('/v1/bookmarks/%s?is_all=true' % id)
+        response = self.client.delete('/api/bookmarks/%s?is_all=true' % id)
         self.assertEqual(response.status_code, 403)
 
     # ToDo: put is allowed by owner user only.
@@ -468,7 +467,7 @@ class APITest(TestCase):
         self.client.logout()
         self.client.login(username=v.username1, password=v.password1)
 
-        r = self.client.get('/v1/bookmarks')
+        r = self.client.get('/api/bookmarks')
         id = json.loads(r.content).get('results')[0].get('id')
 
         payload = {'id': id,
@@ -477,7 +476,7 @@ class APITest(TestCase):
                    'category': v.category0,
                    'description': v.description1,
                    'is_hide': True}
-        response = self.client.put('/v1/bookmarks/%s' % id,
+        response = self.client.put('/api/bookmarks/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
@@ -486,7 +485,7 @@ class APITest(TestCase):
         self.client.logout()
         payload = {'bookmark': v.url0,
                    'tag': v.tag1}
-        response = self.client.post('/v1/bookmark_tags',
+        response = self.client.post('/api/bookmark_tags',
                                     content_type='application/json',
                                     data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
@@ -494,21 +493,21 @@ class APITest(TestCase):
     def test_delete_bookmark_tags_by_anonymous(self):
         self.client.logout()
 
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
 
-        response = self.client.delete('/v1/bookmark_tags/%s' % id)
+        response = self.client.delete('/api/bookmark_tags/%s' % id)
         self.assertEqual(response.status_code, 403)
 
     def test_put_bookmark_tags_by_anonymous(self):
         self.client.logout()
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id,
                    'bookmark': v.url0,
                    'tag': v.tag1}
 
-        response = self.client.put('/v1/bookmark_tags/%s' % id,
+        response = self.client.put('/api/bookmark_tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)
@@ -519,9 +518,9 @@ class APITest(TestCase):
         self.client.logout()
         self.client.login(username=v.username1, password=v.password1)
 
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
-        response = self.client.delete('/v1/bookmark_tags/%s' % id)
+        response = self.client.delete('/api/bookmark_tags/%s' % id)
         self.assertEqual(response.status_code, 403)
 
     # ToDo: put is allowed by owner user only.
@@ -530,12 +529,12 @@ class APITest(TestCase):
         self.client.logout()
         self.client.login(username=v.username1, password=v.password1)
 
-        r = self.client.get('/v1/bookmark_tags')
+        r = self.client.get('/api/bookmark_tags')
         id = json.loads(r.content).get('results')[0].get('id')
         payload = {'id': id,
                    'bookmark': v.url0,
                    'tag': v.tag1}
-        response = self.client.put('/v1/bookmark_tags/%s' % id,
+        response = self.client.put('/api/bookmark_tags/%s' % id,
                                    content_type='application/json',
                                    data=json.dumps(payload))
         self.assertEqual(response.status_code, 403)